requirements.txt
accelerate==1.10.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
filelock==3.13.1
flask==3.1.2
flatbuffers==25.9.23
fsspec==2024.6.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
onnxruntime-gpu==1.20.1
packaging==25.0
pandas==2.3.3
pillow==11.0.0
protobuf==6.32.1
psutil==7.1.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
timm==1.0.20
tokenizers==0.22.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.0
typing-extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
werkzeug==3.1.3